Made smget_area_size() signed.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@692 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/base/misc/smalloc.c
@@ -209,7 +209,7 @@ size_t smget_free_space(struct memnode *mp)
   return count;
 }
 
-size_t smget_area_size(struct memnode *mp, void *ptr)
+int smget_area_size(struct memnode *mp, void *ptr)
 {
   struct memnode *mn;
   for (mn = mp->next; mn; mn = mn->next) {
